% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simml.main.R
\name{der.link}
\alias{der.link}
\title{A subfunction used in estimation}
\usage{
der.link(g.fit, eps = 10^(-6))
}
\arguments{
\item{g.fit}{a \code{mgcv::gam} object}

\item{eps}{a small finite difference used in numerical differentiation.}
}
\description{
This function computes the 1st derivative of the treatment-specific link function with respect to the single index, using finite difference.
}
\seealso{
\code{fit.simml}, \code{simml}
}
